remove long long ints

+++ main.h
@@ -14,7 +14,7 @@ int _printf(const char *format, ...);
 int choice(va_list r, char str);
 int printbin(long int num);
 int printoctal(long int num);
-int printhex(long long int num);
+int printhex(long int num);
 
 /**
  * struct choice - struct to nhold the choice of a person
